+++ Framework-Headers/AccessibilityBundles/UIKit.axbundle/Objc/UIInputSetHostViewAccessibility.h
@@ -14,7 +14,7 @@
 + (Class)safeCategoryBaseClass;
 + (id)safeCategoryTargetClassName;
 - (BOOL)_accessibilityIsScannerGroup;
-- (int)_accessibilityScannerGroupTraits;
+- (unsigned long long)_accessibilityScannerGroupTraits;
 
 @end
