+++ Framework-Headers/PrivateFrameworks/UIAccessibility._framework/Objc/AXRemoteElement+UIAccessibility.h
@@ -10,8 +10,8 @@
 - (BOOL)_accessibilityIsRemoteElement;
 - (id)_ancestorElementThatSupportsActivationAction;
 - (BOOL)accessibilityActivate;
-- (id)accessibilityAttributeValue:(int)arg1;
-- (id)accessibilityAttributeValue:(int)arg1 forParameter:(id)arg2;
+- (id)accessibilityAttributeValue:(long long)arg1;
+- (id)accessibilityAttributeValue:(long long)arg1 forParameter:(id)arg2;
 - (BOOL)accessibilityViewIsModal;
 - (void)platformCleanup;
 @end
